bsns页面修改

+++ src/main/java/cn/edu/swust/service/BusinessService.java
@@ -23,6 +23,7 @@ import cn.edu.swust.query.RolePermissionQuery;
 import cn.edu.swust.query.StepQuery;
 import cn.edu.swust.query.UserQuery;
 import cn.edu.swust.util.LogHelper;
+import cn.edu.swust.util.ObjectConvert;
 
 @Service
 public class BusinessService {
@@ -38,48 +39,97 @@ public class BusinessService {
 	private DeclareStepDao declareStepDao;
 	
 	public List<Business> getBusinessList(BusinessQuery query){
-		if(query!=null) {
+//		if(query!=null) {
+//			List<Business> list = businessDao.queryList(query);
+//			List<Business> businesses = new ArrayList<>();
+//			try {
+//				for(Business bsn : list) {
+//					UserQuery userQuery = new UserQuery();
+//					userQuery.setUserId(bsn.getCreateUserId());
+//					bsn.setCreateUser(userDao.query(userQuery));
+//					businesses.add(bsn);
+//				}
+//				
+//				return businesses;
+//			} catch (Exception e) {
+//				LogHelper.logError(e);
+//			}
+//			
+//		}
+		List<Business> businesses = new ArrayList<>();
+		try {
 			List<Business> list = businessDao.queryList(query);
-			List<Business> businesses = new ArrayList<>();
-			try {
-				for(Business bsn : list) {
-					UserQuery userQuery = new UserQuery();
-					userQuery.setUserId(bsn.getCreateUserId());
-					bsn.setCreateUser(userDao.query(userQuery));
-					businesses.add(bsn);
-				}
+			for(Business bsn : list) {
+				UserQuery userQuery = new UserQuery();
+				userQuery.setUserId(bsn.getCreateUserId());
 				
-				return businesses;
-			} catch (Exception e) {
-				LogHelper.logError(e);
+				bsn.setCreateUser(userDao.query(userQuery));
+				
+				StepQuery stepQuery = new StepQuery();
+				stepQuery.setBusinessId(bsn.getBusinessId());
+				
+				bsn.setSteps(stepDao.queryList(stepQuery));
+				
+				businesses.add(bsn);
 			}
-			
+		} catch (Exception e) {
+			LogHelper.logError(e);
 		}
-		return null;
+		return businesses;
 	}
 	/***
 	 * 添加项目
 	 * @param business  项目信息
 	 * @param steps   项目审批过程 list
-	 * @return
+	 * @return  插入Id
 	 */
 	@Transactional
 	public int insertBusiness(Business business,List<Step> steps) {
-		int ret = 0;
+		int sltId = 0;
 		try {
+			if(business.getCreateUserId()==null||business.getCreateUserId()==""||steps==null||steps.size()==0) {
+				return sltId;
+			}
+			
 			businessDao.setPrimaryValue(1);
-			ret = businessDao.insertOneSelective(business);
-			stepDao.setPrimaryValue(1);
-			stepDao.batchInsert(steps);
+			sltId = businessDao.insertOneSelective(business);
+			
+			if(sltId != 0) {
+				for(Step step : steps) {
+					step.setBusinessId(sltId);
+				}
+				stepDao.setPrimaryValue(1);
+				stepDao.batchInsert(steps);
+			}
+			
 		} catch (Exception e) {
 			LogHelper.logError(e);
 			TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
-			ret = 0;
+			sltId = 0;
 		}
-		return ret;
+		return sltId;
 	}
 	/**
-	 * 删除项目信息 
+	 * 批量删除  
+	 * @param ids
+	 * @return  0是全部失败  1是部分失败  2是全部成功
+	 */
+	public int batchDelete(int[] ids) {
+		int successCount = 0;
+		if(ids!=null && ids.length>0) {
+			BusinessQuery query = new BusinessQuery();
+			for(int id : ids) {
+				query.setBusinessId(id);
+				successCount += delete(query)>0?1:0;
+			}
+		}
+		if(successCount==0) {
+			return 0;
+		}
+		return successCount==ids.length?2:1;
+	}
+	/**
+	 * 删除项目信息 删除business表数据、删除step表数据、已经申报的项目不能删除
 	 * @param query
 	 * @return 影响行数
 	 */
@@ -88,31 +138,22 @@ public class BusinessService {
 		int ret = 0;
 		try {
 			//todo 批量删除bug  修改删除项目的逻辑
-			//逻辑：删除business表数据、删除step表数据、已经申报的项目不能删除
-			StepQuery stepQuery = new StepQuery();
+			
 			DeclareBusinessQuery dclrBsnsQuery = new DeclareBusinessQuery();
-			DeclareStepQuery dclrDlQuery = new DeclareStepQuery();
-			if(query.getBusinessId()!=0) {
-				stepQuery.setBusinessId(query.getBusinessId());
-//				dclrBsnsQuery.setBusinessId(query.getBusinessId());
-//				dclrDlQuery.setDeclareBusinessId(query.getBusinessId());
+			ObjectConvert.obj2Obj(query, dclrBsnsQuery);
+
+			boolean isDeclare = declareBusinessDao.count(dclrBsnsQuery)==0?false:true;
+
+			if(!isDeclare) {
+				ret = businessDao.delete(query);
+				StepQuery stepQuery = new StepQuery();
+				ObjectConvert.obj2Obj(query, stepQuery);
+				stepDao.delete(stepQuery);
 			}
-			if(query.getBusinessIds()!=null && query.getBusinessIds().length>0) {
-				stepQuery.setBusinessIds(query.getBusinessIds());
-//				dclrBsnsQuery.setBusinessIds(query.getBusinessIds());
-//				dclrDlQuery.setDeclareBusinessIds(query.getBusinessIds());
-			}
-			
-//			declarestepDao.delete(dclrDlQuery);
-//			declareBusinessDao.delete(dclrBsnsQuery);
-			stepDao.delete(stepQuery);
-			
-			ret = businessDao.delete(query);
 		}catch (Exception ex) {
 			LogHelper.logError(ex);
 			TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
 			ret = 0;
-			
 		}
 		return ret;
 	}
